requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
chromadb>=0.5.3
pydantic>=2.7.0
flask>=3.0
flask-cors>=4.0
tk
tk-tools>=0.1.0